+++ AppboyProject/AppboyProject.js
@@ -6,7 +6,8 @@ import {
   View,
   TouchableHighlight,
   Linking,
-  Alert
+  Alert,
+  TextInput
 } from 'react-native';
 
 const ReactAppboy = require('react-native-appboy-sdk');
@@ -14,8 +15,13 @@ const ReactAppboy = require('react-native-appboy-sdk');
 class AppboyProject extends Component {
   constructor(props) {
     super(props);
-    this.state = {};
+    this.state = {
+      userIdText : 'theAppboyTestUser',
+      customEventText : ''
+    };
     this._updateCardCount = this._updateCardCount.bind(this);
+    this._changeUserPress = this._changeUserPress.bind(this);
+    this._logCustomEventPress = this._logCustomEventPress.bind(this);
   }
 
   componentDidMount() {
@@ -82,14 +88,26 @@ class AppboyProject extends Component {
   render() {
     return (
       <View style={styles.container}>
-        <TouchableHighlight
-          onPress={this._changeUserPress}>
-          <Text>Change User</Text>
-        </TouchableHighlight>
-        <TouchableHighlight
-          onPress={this._logCustomEventPress}>
-          <Text>Log Custom Event</Text>
-        </TouchableHighlight>
+        <View style={styles.row}>
+          <TextInput
+            style={styles.textInput}
+            onChangeText={(userIdText) => this.setState({userIdText})}
+            value={this.state.userIdText}
+            />
+          <TouchableHighlight
+            onPress={this._changeUserPress}>
+            <Text>Set User ID</Text>
+          </TouchableHighlight>
+        </View>
+        <View style={styles.row}>
+          <TextInput
+            style={styles.textInput}
+            onChangeText={(customEventText) => this.setState({customEventText})}/>
+          <TouchableHighlight
+            onPress={this._logCustomEventPress}>
+            <Text>Log Custom Event</Text>
+          </TouchableHighlight>
+        </View>
         <TouchableHighlight
           onPress={this._logPurchasePress}>
           <Text>Log Purchase</Text>
@@ -105,10 +123,6 @@ class AppboyProject extends Component {
         <TouchableHighlight
           onPress={this._logUserPropertiesPress}>
           <Text>Set User Properties</Text>
-        </TouchableHighlight>
-        <TouchableHighlight
-          onPress={this._launchNewsFeedPress}>
-          <Text>Launch News Feed</Text>
         </TouchableHighlight>
         <TouchableHighlight
           onPress={this._launchFeedbackPress}>
@@ -138,17 +152,29 @@ class AppboyProject extends Component {
           onPress={this._setFacebookData}>
           <Text>Set Facebook Data</Text>
         </TouchableHighlight>
+        <TouchableHighlight
+          onPress={this._launchNewsFeedPress}>
+          <Text>Launch News Feed</Text>
+        </TouchableHighlight>
         <TouchableHighlight onPress={this._updateCardCount}>
           <Text>Unread Cards (Click to Refresh): {this.state.unreadCardCount} / {this.state.cardCount}</Text>
+        </TouchableHighlight>
+        <TouchableHighlight
+          onPress={this._requestFeedRefresh}>
+          <Text>Request Feed Refresh</Text>
+        </TouchableHighlight>
+        <TouchableHighlight
+          onPress={this._requestImmediateDataFlush}>
+          <Text>Request Immediate Data Flush</Text>
         </TouchableHighlight>
       </View>
     );
   }
   _changeUserPress(event) {
-    ReactAppboy.changeUser('theAppboyTestUser');
+    ReactAppboy.changeUser(this.state.userIdText);
   }
   _logCustomEventPress(event) {
-    ReactAppboy.logCustomEvent('reactCustomEvent', {'p1': 'p2'});
+    ReactAppboy.logCustomEvent(this.state.customEventText, {'p1': 'p2'});
   }
   _logPurchasePress(event) {
     ReactAppboy.logPurchase('reactProductIdentifier', '1.2', 'USD', 2, {'pp1': 'pp2'});
@@ -238,6 +264,14 @@ class AppboyProject extends Component {
     ];
     ReactAppboy.setFacebookData(profile, 500, likes);
   }
+
+  _requestFeedRefresh(event) {
+    ReactAppboy.requestFeedRefresh();
+  }
+
+  _requestImmediateDataFlush(event) {
+    ReactAppboy.requestImmediateDataFlush();
+  }
 }
 
 const styles = StyleSheet.create({
@@ -247,15 +281,18 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     backgroundColor: '#F5FCFF'
   },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10
+  textInput: {
+    height: 40,
+    width: 150,
+    borderColor: 'gray',
+    borderWidth: .5,
+    paddingLeft: 5,
+    marginLeft: 5,
+    fontSize: 14
   },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5
+  row: {
+    flexDirection: 'row',
+    alignItems: 'center'
   }
 });